requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.12.0=pyhd8ed1ab_0
aiohttp=3.7.4=py38h294d835_0
argon2-cffi=20.1.0=py38h2bbff1b_1
async-timeout=3.0.1=py_1000
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.9.3=pyhb0f4dca_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blinker=1.4=py_1
brotlipy=0.7.0=py38h294d835_1001
ca-certificates=2021.5.30=h5b45459_0
cachetools=4.2.2=pyhd8ed1ab_0
catboost=0.25.1=py38haa244fe_0
certifi=2021.5.30=py38haa244fe_0
cffi=1.14.5=py38hcd4344a_0
chardet=4.0.0=py38haa244fe_1
click=8.0.1=py38haa244fe_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.7=py38hd7da0ea_0
cudatoolkit=11.0.3=h3f58a73_8
cycler=0.10.0=py_2
cytoolz=0.11.0=py38he774522_0
dask-core=2.30.0=py_0
dataclasses=0.8=pyhc8e2a94_1
dataframe_image=0.1.1=py_0
decorator=5.0.7=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py38_0
freeglut=3.2.1=h0e60522_2
freetype=2.10.4=h546665d_1
fsspec=2021.6.0=pyhd8ed1ab_0
future=0.18.2=py38haa244fe_3
google-auth=1.30.0=pyh44b312d_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.38.0=py38he5377a8_0
hdf5=1.10.4=nompi_hcc15c50_1106
icu=68.1=h0e60522_0
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=py_0
importlib-metadata=3.10.0=py38haa95532_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2021.2.0=haa95532_616
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.22.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3deb0d_0
jasper=2.0.14=h77af90b_2
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_contrib_core=0.3.3=py_2
jupyter_contrib_nbextensions=0.5.1=pyhd8ed1ab_2
jupyter_core=4.7.1=py38haa95532_0
jupyter_highlight_selected_word=0.2.0=py38haa244fe_1002
jupyter_latex_envs=1.4.6=pyhd8ed1ab_1002
jupyter_nbextensions_configurator=0.4.1=py38haa244fe_2
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
kiwisolver=1.3.1=py38hbd9d945_1
lcms2=2.12=h2a16943_0
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=5_hd5c7e75_netlib
libclang=11.1.0=default_h5c34c98_1
libiconv=1.16=he774522_0
liblapack=3.9.0=5_hd5c7e75_netlib
liblapacke=3.9.0=5_hd5c7e75_netlib
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h1d00b33_2
libprotobuf=3.17.2=h7755175_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.2.0=hc10be44_1
libuv=1.41.0=h8ffe710_0
libwebp-base=1.2.0=h8ffe710_2
libxml2=2.9.12=hf5bbc77_0
libxslt=1.1.33=h65864e5_2
llvmlite=0.36.0=py38h57a6900_0
lxml=4.6.3=py38h292cb97_0
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=1.1.1=py38he774522_0
matplotlib=3.4.2=py38haa244fe_0
matplotlib-base=3.4.2=py38heae8d8c_0
mistune=0.8.4=py38he774522_1000
mkl=2021.2.0=haa95532_296
mkl-service=2.3.0=py38h2bbff1b_1
mkl_fft=1.3.0=py38h277e83a_2
mkl_random=1.2.1=py38hf11a4ad_2
msys2-conda-epoch=20160418=1
multidict=5.1.0=py38h294d835_1
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=py_0
ninja=1.10.2=h5362a0b_0
notebook=6.3.0=py38haa95532_0
numba=0.53.1=py38hd0eb412_0
numpy=1.20.1=py38h34a8a5c_0
numpy-base=1.20.1=py38haf7ebc8_0
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.0.1=py38h2a7c758_0
openjpeg=2.4.0=h48faf41_0
openssl=1.1.1k=h8ffe710_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.4=py38hd77b12b_0
pandas-profiling=1.4.1=0
pandoc=2.12=haa95532_0
pandocfilters=1.4.3=py38haa95532_1
parso=0.8.2=pyhd3eb1b0_0
patsy=0.5.1=py_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.2=py38h9273828_1
pip=21.0.1=py38haa95532_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.17.2=py38h885f38d_0
py-opencv=4.0.1=py38he44ac1e_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_2
pydeprecate=0.3.0=pyhd8ed1ab_0
pygments=2.8.1=pyhd3eb1b0_0
pyjwt=2.1.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.12.3=py38haa244fe_7
pyqt-impl=5.12.3=py38h885f38d_7
pyqt5-sip=4.19.18=py38h885f38d_7
pyqtchart=5.12=py38h885f38d_7
pyqtwebengine=5.12.1=py38h885f38d_7
pyrsistent=0.17.3=py38he774522_0
pysocks=1.7.1=py38haa244fe_3
python=3.8.8=hdbf39b2_5
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.7.0=py3.8_cuda110_cudnn8_0
pytorch-lightning=1.3.5=pyhd8ed1ab_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py38he774522_2
pywin32=227=py38he774522_1
pywinpty=0.5.7=py38_0
pyyaml=5.3.1=py38he774522_0
pyzmq=20.0.0=py38hd77b12b_1
qt=5.12.9=h5909a2a_4
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
scikit-image=0.17.2=py38h1e1f486_0
scikit-learn=0.24.2=py38h5d5d464_0
scipy=1.6.3=py38he847743_0
seaborn=0.11.1=hd8ed1ab_1
seaborn-base=0.11.1=pyhd8ed1ab_1
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py38haa95532_0
shap=0.39.0=py38h60cbd38_0
six=1.15.0=py38haa95532_0
slicer=0.0.7=pyhd8ed1ab_0
soupsieve=2.0.1=py_1
sqlite=3.35.4=h2bbff1b_0
statsmodels=0.12.2=py38h347fdf6_0
tensorboard=2.4.1=pyhd8ed1ab_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
terminado=0.9.4=py38haa95532_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py38h8c2d366_2
tk=8.6.10=h8ffe710_1
toolz=0.11.1=py_0
torchmetrics=0.3.2=pyhd8ed1ab_0
torchvision=0.8.1=py38_cu110
tornado=6.1=py38h2bbff1b_0
tqdm=4.60.0=pyhd8ed1ab_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38haa244fe_4
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38_0
winpty=0.4.3=4
xz=5.2.5=h62dcd97_1
yaml=0.1.7=vc14h4cb57cf_1
yarl=1.6.3=py38h294d835_1
zeromq=4.3.3=ha925a31_3
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_1010
zstd=1.4.9=h6255e5f_0
